requirements.txt
dict2obj
discord.py
requests

aws-lambda-typing == 1.0.3
boto3
youtube_dl
ffmpeg
nest_asyncio

black
flake8
autopep8
isort

Flask
mypy
beautifulsoup4
pytest
pytest-mock
pytest-asyncio
pytest-env
